+++ tests_dawgit/test_snapshot_backup.py
@@ -1,3 +1,4 @@
+import ui_strings
 import os
 os.environ["DAWGIT_TEST_MODE"] = "1"